extract refactored behaviour to a different spec file
tt
