+++ script/ci/ordered_tests.txt
@@ -1,4 +1,5 @@
 spec/models/user_part_validation_and_authentication_spec.rb
+spec/models/user_part_refactored_behaviour_spec.rb
 spec/models/user_part_app_spec.rb
 spec/models/carto/user_migration_rollback_spec.rb
 spec/requests/carto/api/visualizations_controller_spec_part_2.rb
